+++ src/collective/chimpfeed/schema.py
@@ -1,4 +1,5 @@
 from zope.interface import implements
+from zope.interface import alsoProvides
 from zope.component import adapts
 from archetypes.schemaextender.interfaces import ISchemaExtender
 from archetypes.schemaextender.field import ExtensionField
@@ -6,15 +7,60 @@ from plone.indexer.decorator import indexer
 
 from Products.Archetypes import atapi
 from Products.Archetypes.interfaces import IBaseContent
+from DateTime import DateTime
 
-from collective.chimpfeed import MessageFactory as _
 from collective.chimpfeed.permissions import MODERATE_PERMISSION
+from collective.chimpfeed.interfaces import IFeedControl
+
+
+try:
+    from plone.app.dexterity.behaviors.metadata import Categorization
+    from plone.app.dexterity.behaviors.metadata import DCFieldProperty
+except ImportError:
+    pass
+else:
+    class FeedControl(Categorization):
+        feeds = DCFieldProperty(IFeedControl['feeds'])
+        feedSchedule = DCFieldProperty(IFeedControl['feedSchedule'])
+        feedModerate = DCFieldProperty(IFeedControl['feedModerate'])
+
+
+try:
+    from plone.autoform.interfaces import IFormFieldProvider
+except ImportError:
+    pass
+else:
+    alsoProvides(IFeedControl, IFormFieldProvider)
 
 
 @indexer(IBaseContent)
-def feed_indexer(context):
+def at_feed_indexer(context):
     feeds = context.getField('feeds').get(context)
-    return [unicode(feed, 'utf-8') for feed in feeds]
+    return tuple(unicode(feed, 'utf-8') for feed in feeds)
+
+
+try:
+    from plone.dexterity.interfaces import IDexterityContent
+except ImportError:
+    pass
+else:
+    @indexer(IDexterityContent)
+    def dx_schedule_indexer(context):
+        date = context.feedSchedule
+        return DateTime(
+            date.year,
+            date.month,
+            date.day
+            )
+
+    @indexer(IDexterityContent)
+    def dx_feed_indexer(context):
+        if IFeedControl.providedBy(context):
+            return tuple(context.feeds)
+
+        # To-Do: It seems that we can't check for the behavior in this
+        # way.
+        return getattr(context, "feeds", None)
 
 
 class ScheduleField(ExtensionField, atapi.DateTimeField):
@@ -49,12 +95,10 @@ class FeedExtender(object):
             multivalued=1,
             schemata="settings",
             widget=atapi.MultiSelectionWidget(
-                label=_("Feeds"),
-                description=_(u"If you want this content item published "
-                              u"to a news feed, select one or more from "
-                              u"the list below."),
+                label=IFeedControl['feeds'].title,
+                description=IFeedControl['feeds'].description,
                 ),
-            vocabulary_factory="collective.chimpfeed.vocabularies.Feeds",
+            vocabulary_factory=IFeedControl['feeds'].value_type.vocabulary,
             ),
 
         ScheduleField(
@@ -67,10 +111,8 @@ class FeedExtender(object):
                 show_hm=False,
                 starting_year=2012,
                 ending_year=2015,
-                label=_(u'Schedule feed date'),
-                description=_(u'The item will be scheduled to be included '
-                              u'in feeds from this date. Note that this '
-                              u'is subject to editor moderation.'),
+                label=IFeedControl['feedSchedule'].title,
+                description=IFeedControl['feedSchedule'].description
                 ),
             ),
 
@@ -81,8 +123,8 @@ class FeedExtender(object):
             enforceVocabulary=1,
             write_permission=MODERATE_PERMISSION,
             widget=atapi.BooleanWidget(
-                label=_(u'Approve schedule'),
-                description=_(u'Select this option to approve schedule.'),
+                label=IFeedControl['feedModerate'].title,
+                description=IFeedControl['feedModerate'].description,
                 ),
             ),
         )